+++ sdk/src/test/java/com/finbourne/lusid/tutorials/ibor/PortfoliosTest.java
@@ -13,6 +13,7 @@ import org.junit.BeforeClass;
 import org.junit.Test;
 import org.junit.Ignore;
 
+import java.math.BigDecimal;
 import java.time.OffsetDateTime;
 import java.time.ZoneOffset;
 import java.util.*;
@@ -157,11 +158,11 @@ public class PortfoliosTest {
                             { put(LUSID_INSTRUMENT_IDENTIFIER, instrumentIds.get(0)); }
                         }
                 )
-                .totalConsideration(new CurrencyAndAmount().currency("GBP").amount(1230.0))
+                .totalConsideration(new CurrencyAndAmount().currency("GBP").amount(BigDecimal.valueOf(1230.0)))
                 .transactionDate(effectiveDate.toString())
                 .settlementDate(effectiveDate.toString())
-                .units(100.0)
-                .transactionPrice(new TransactionPrice().price(12.3))
+                .units(BigDecimal.valueOf(100.0))
+                .transactionPrice(new TransactionPrice().price(BigDecimal.valueOf(12.3)))
                 .source("Broker");
 
         //    Add the transaction to the portfolio
@@ -229,11 +230,11 @@ public class PortfoliosTest {
                             { put(LUSID_INSTRUMENT_IDENTIFIER, instrumentIds.get(0)); }
                         }
                  )
-                .totalConsideration(new CurrencyAndAmount().currency("GBP").amount(1230.0))
+                .totalConsideration(new CurrencyAndAmount().currency("GBP").amount(BigDecimal.valueOf(1230.0)))
                 .transactionDate(effectiveDate.toString())
                 .settlementDate(effectiveDate.toString())
-                .units(100.0)
-                .transactionPrice(new TransactionPrice().price(12.3))
+                .units(BigDecimal.valueOf(100.0))
+                .transactionPrice(new TransactionPrice().price(BigDecimal.valueOf(12.3)))
                 .source("Custodian")
                 .properties(properties);
